+++ pages/src/components/Page/Page.js
@@ -2,7 +2,7 @@ var Page = Backbone.View.extend({
     el: '.main',
 
     render: function() {
-        this.$('aside .page').parent().addClass('active');
+        $('aside .page').parent().addClass('active');
 
         var view = this;
         $.get('src/components/Page/page.html', function(tpl) {