+++ pages/src/components/Tooltip/Tooltip.js
@@ -2,7 +2,7 @@ var Tooltip = Backbone.View.extend({
     el: '.main',
 
     render: function() {
-        $('aside .tooltip').parent().addClass('active');
+        $('aside .tooltips').parent().addClass('active');
 
         var view = this;
         $.get('src/components/Tooltip/tooltip.html', function(tpl) {